requirements/dev.txt
-r base.txt
coverage==4.0.1
django-nose==1.4.2
ipython==4.0.0
mock==1.3.0
